merge

+++ app/src/main/java/pattern/builder/experiment/Builder.java
@@ -1,8 +1,13 @@
+/**
+ * Builder Interface
+ */
+
 package pattern.builder.experiment;
 
 public interface Builder {
-    public void addSolute(double soluteAmount);
-    public void addSolvent(double solventAmount);
-    public void abandonSolution(double solutionAmount);
+
+    public void addSolute(float soluteAmount);
+    public void addSolvent(float solventAmount);
+    public void abandonSolution(float solutionAmount);
     public Object getResult();
 }
