+++ app/src/main/java/pattern/builder/experiment/SaltWater.java
@@ -1,9 +1,10 @@
 package pattern.builder.experiment;
 
-public class SaltWater{
-    public double salt;
-    public double water;
-    public SaltWater(double water,double salt){
+public class SaltWater {
+    public float salt;
+    public float water;
+
+    public SaltWater(float water, float salt) {
         this.salt = salt;
         this.water = water;
     }
