+++ app/src/main/java/pattern/builder/experiment/SaltWaterBuilder.java
@@ -1,25 +1,30 @@
 package pattern.builder.experiment;
 
-public class SaltWaterBuilder implements Builder{
+public class SaltWaterBuilder implements Builder {
     private SaltWater saltWater;
-    public SaltWaterBuilder(){
-        this.saltWater = new SaltWater(0,0);
+
+    public SaltWaterBuilder() {
+        this.saltWater = new SaltWater(0, 0);
     }
-    public void addSolute(double saltAmount){
+
+    public void addSolute(float saltAmount) {
         saltWater.salt += saltAmount;
     }
-    public void addSolvent(double waterAmount){
+
+    public void addSolvent(float waterAmount) {
         saltWater.water += waterAmount;
     }
-    public void abandonSolution(double saltWaterAmount){
-        double saltDelta = saltWaterAmount *
+
+    public void abandonSolution(float saltWaterAmount) {
+        float saltDelta = saltWaterAmount *
                 (saltWater.salt / (saltWater.salt + saltWater.water));
-        double waterDelta = saltWaterAmount *
+        float waterDelta = saltWaterAmount *
                 (saltWater.water / (saltWater.salt + saltWater.water));
         saltWater.salt -= saltDelta;
         saltWater.water -= waterDelta;
     }
-    public Object getResult(){
+
+    public Object getResult() {
         return this.saltWater;
     }
 }
